+++ src/common/app.release.ts
@@ -27,7 +27,7 @@ export const Release = {
 
   // this is here to trigger revalidation of data, e.g. models refresh
   Monotonics: {
-    Aix: 6,
+    Aix: 8,
     NewsVersion: 191,
   },