+++ src/modules/aix/server/dispatch/chatGenerate/chatGenerate.dispatch.ts
@@ -47,7 +47,7 @@ export function createChatGenerateDispatch(access: AixAPI_Access, model: AixAPI_
     case 'gemini':
       return {
         request: {
-          ...geminiAccess(access, model.id, streaming ? GeminiWire_API_Generate_Content.streamingPostPath : GeminiWire_API_Generate_Content.postPath),
+          ...geminiAccess(access, model.id, streaming ? GeminiWire_API_Generate_Content.streamingPostPath : GeminiWire_API_Generate_Content.postPath, !!model.vndGeminiShowThoughts),
           body: aixToGeminiGenerateContent(model, chatGenerate, access.minSafetyLevel, false, streaming),
         },
         demuxerFormat: streaming ? 'sse' : null,