+++ src/modules/llms/models-modal/LLMParametersEditor.tsx
@@ -23,6 +23,7 @@ export function LLMParametersEditor(props: {
   // consts
   maxOutputTokens: number | null,
   parameterSpecs: DModelParameterSpec<DModelParameterId>[],
+  parameterOmitTemperature?: boolean,
   baselineParameters: DModelParameterValues,
 
   // value and onChange for the parameters
@@ -38,11 +39,12 @@ export function LLMParametersEditor(props: {
   const allParameters = getAllModelParameterValues(baselineParameters, parameters);
 
   // derived state
-  const llmTemperature = allParameters.llmTemperature ?? FALLBACK_LLM_PARAM_TEMPERATURE;
+  const llmTemperature: number | null = allParameters.llmTemperature === undefined ? FALLBACK_LLM_PARAM_TEMPERATURE : allParameters.llmTemperature;
   const llmResponseTokens = allParameters.llmResponseTokens ?? FALLBACK_LLM_PARAM_RESPONSE_TOKENS;
+  const llmVndGeminiShowThoughts = allParameters.llmVndGeminiShowThoughts;
   const llmVndOaiReasoningEffort = allParameters.llmVndOaiReasoningEffort;
   const llmVndOaiRestoreMarkdown = !!allParameters.llmVndOaiRestoreMarkdown;
-  const tempAboveOne = llmTemperature > 1;
+  const tempAboveOne = llmTemperature !== null && llmTemperature > 1;
 
   // more state (here because the initial state depends on props)
   const [overheat, setOverheat] = React.useState(tempAboveOne);
@@ -60,6 +62,9 @@ export function LLMParametersEditor(props: {
   }, [onChangeParameter, overheat, tempAboveOne]);
 
 
+  // find the gemini show thoughts parameter spec
+  const paramSpecGeminiShowThoughts = parameterSpecs?.find(p => p.paramId === 'llmVndGeminiShowThoughts') as DModelParameterSpec<'llmVndGeminiShowThoughts'> | undefined;
+
   // find the reasoning effort parameter spec
   // NOTE: optimization: this controls both the 'reasoning effort' and 'restore markdown' UI settings
   const paramSpecReasoningEffort = parameterSpecs?.find(p => p.paramId === 'llmVndOaiReasoningEffort') as DModelParameterSpec<'llmVndOaiReasoningEffort'> | undefined;
@@ -70,7 +75,8 @@ export function LLMParametersEditor(props: {
 
     <FormSliderControl
       title='Temperature' ariaLabel='Model Temperature'
-      description={llmTemperature < 0.33 ? 'More strict' : llmTemperature > 1 ? 'Extra hot ♨️' : llmTemperature > 0.67 ? 'Larger freedom' : 'Creativity'}
+      description={llmTemperature === null ? 'Unsupported' : llmTemperature < 0.33 ? 'More strict' : llmTemperature > 1 ? 'Extra hot ♨️' : llmTemperature > 0.67 ? 'Larger freedom' : 'Creativity'}
+      disabled={props.parameterOmitTemperature}
       min={0} max={overheat ? 2 : 1} step={0.1} defaultValue={0.5}
       valueLabelDisplay={parameters?.llmTemperature !== undefined ? 'on' : 'auto'}
       value={llmTemperature}
@@ -100,6 +106,15 @@ export function LLMParametersEditor(props: {
       </Box>
     ) : (
       <InlineError error='Max Output Tokens: Token computations are disabled because this model does not declare the context window size.' />
+    )}
+
+    {paramSpecGeminiShowThoughts && (
+      <FormSwitchControl
+        title='Show Chain of Thought'
+        description={`Displays Gemini\'s reasoning process`}
+        checked={!!llmVndGeminiShowThoughts}
+        onChange={checked => onChangeParameter({ llmVndGeminiShowThoughts: checked })}
+      />
     )}
 
     {paramSpecReasoningEffort && (