+++ src/modules/youtube/YouTubeURLInput.tsx
@@ -1,11 +1,15 @@
 import * as React from 'react';
+import { useQuery } from '@tanstack/react-query';
 
 import type { SxProps } from '@mui/joy/styles/types';
 import { Box, Button, Input } from '@mui/joy';
 import YouTubeIcon from '@mui/icons-material/YouTube';
 
+import { InlineError } from '~/common/components/InlineError';
+
+import type { YouTubeVideoData } from './youtube.types';
 import { extractYoutubeVideoIDFromURL } from './youtube.utils';
-import { useYouTubeTranscript, YTVideoTranscript } from './useYouTubeTranscript';
+import { youTubeGetVideoData } from './useYouTubeTranscript';
 
 
 interface YouTubeURLInputProps {
@@ -17,53 +21,63 @@ export const YouTubeURLInput: React.FC<YouTubeURLInputProps> = ({ onSubmit, sx }
 
   // state
   const [url, setUrl] = React.useState('');
-  const [submitFlag, setSubmitFlag] = React.useState(false);
 
-  const videoID = extractYoutubeVideoIDFromURL(url);
+  // derived state
+  const validVideoId = extractYoutubeVideoIDFromURL(url);
 
-  // Callback function to handle new transcript
-  const handleNewTranscript = React.useCallback((newTranscript: YTVideoTranscript) => {
-    onSubmit(newTranscript.transcript); // Pass the transcript text to the onSubmit handler
-    setSubmitFlag(false); // Reset submit flag after handling
-  }, [onSubmit]);
+  // query
+  const { isFetching: isTranscriptFetching, refetch: refetchTranscript, isError, error, isSuccess } = useQuery({
+    enabled: false,
+    queryKey: ['videoData', validVideoId],
+    queryFn: async (): Promise<YouTubeVideoData> => youTubeGetVideoData(validVideoId!),
+    staleTime: Infinity,
+  });
 
-  const { isFetching: isTranscriptFetching, isError, error } = useYouTubeTranscript(videoID && submitFlag ? videoID : null, handleNewTranscript);
 
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  const handleChange = React.useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     setUrl(event.target.value);
-  };
+  }, [setUrl]);
 
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+  const handleSubmit = React.useCallback(async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault(); // Prevent form from causing a page reload
-    setSubmitFlag(true); // Set flag to indicate a submit action
-  };
+
+    const result = await refetchTranscript();
+    if (result.data)
+      onSubmit(result.data.transcript);
+
+  }, [onSubmit, refetchTranscript]);
+
 
   return (
-    <Box sx={{ mb: 1, ...sx }}>
+    <Box sx={sx}>
       <form onSubmit={handleSubmit}>
-        <Input
-          required
-          type='url'
-          fullWidth
-          disabled={isTranscriptFetching}
-          variant='outlined'
-          placeholder='Enter YouTube Video URL'
-          value={url}
-          onChange={handleChange}
-          startDecorator={<YouTubeIcon sx={{ color: '#f00' }} />}
-          sx={{ mb: 1.5, backgroundColor: 'background.popup' }}
-        />
-        <Button
-          type='submit'
-          variant='solid'
-          disabled={isTranscriptFetching || !url}
-          loading={isTranscriptFetching}
-          sx={{ minWidth: 140 }}
-        >
-          Get Transcript
-        </Button>
-        {isError && <div>Error fetching transcript. Please try again. ${JSON.stringify(error)}</div>}
+        <Box sx={{ display: 'flex', gap: 1, alignItems: 'center' }}>
+          <Input
+            required
+            type='url'
+            fullWidth
+            disabled={isTranscriptFetching}
+            variant='outlined'
+            placeholder='Enter YouTube Video URL'
+            value={url}
+            onChange={handleChange}
+            startDecorator={<YouTubeIcon sx={{ color: '#f00' }} />}
+            sx={{ backgroundColor: 'background.popup' }}
+          />
+          <Button
+            type='submit'
+            color={isSuccess ? 'success' : isTranscriptFetching ? 'neutral' : 'primary'}
+            variant={isSuccess ? 'soft' : !validVideoId ? 'outlined' : 'solid'}
+            disabled={!validVideoId}
+            loading={isTranscriptFetching}
+            sx={{ minWidth: 140 }}
+            // endDecorator={isSuccess ? '✓' : undefined}
+          >
+            Get Transcript
+          </Button>
+        </Box>
       </form>
+      {isError && <InlineError error={`Error fetching transcript. Please try again. ${error.message}`}></InlineError>}
     </Box>
   );
 };